Moved the whitespace preservation handling to XMLParser so that it gets done during the parsing, not at the runtime.
I think it's more natural way of handling whitespaces --- it preserves
the immutability of Script and avoids unnecessary downcasting
(how does a tag knows how to trim the whitespace inside the body
'Script' object?), and it reduces the runtime overhead.

This change also works better with push toward doing more in the Script
class and avoid Tag instances.